update model class
